Move Customer PayPal tests into own file
What
===
Move Customer PayPal tests into own file.

Why
===
As we add more payment methods the customer tests are going to be pretty
big. Lets start keeping tests for each payment method in separate files.
Transaction tests already do this.
